+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Functional changes:
 - Updated NeoForge dependencies
 - "Primary Spell" now show up as "Main Spell" for better clarity
 - Fix Power Shot does not apply Mark at close ranges #133
+- Add "Missing target" spell casting failure reason, for some spells
 
 # 1.8.2